requirements.txt
Django==1.8
django-crispy-forms==1.6.0
gunicorn==19.6.0
httplib2==0.9.2
lxml==3.6.4
oauth2==1.9.0.post1
Pillow==3.3.1
psycopg2==2.6.1